+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Field `default_encoding` added to the `parquet_encode` processor.
+
 ## 4.10.0 - 2022-10-26
 
 ### Added